# DataBase/apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: DataBase Gateway
  version: 0.1.0
  description: Thin API facade for the DataBase ecosystem.
servers:
  - url: http://127.0.0.1:18090
components:
  securitySchemes:
    DataBaseApiKey:
      type: apiKey
      in: header
      name: X-DataBase-Api-Key
      description: Optional. Required only when DATABASE_GATEWAY_AUTH_REQUIRED=true.
  schemas:
    RequestId:
      type: string
      description: Per-request correlation id returned in X-Request-Id.
...
    SearchResult:
      type: object
      required:
        - document_id
        - source_table
        - source_id
        - source
        - title
        - privacy_level
        - chunk_index
        - snippet
      properties:
        document_id:
          type: string
        source_table:
          type: string
        source_id:
          type: string
        source:
          type:
            - string
            - "null"
        title:
          type:
            - string
            - "null"
        privacy_level:
          type: string
        chunk_index:
          type: integer
        snippet:
          type: string
        score:
          type: number
        provider:
          type: string
    SearchResponse:
      type: object
      required:
        - query
        - count
        - results
        - requestId
      properties:
        query:
          type: string
        count:
          type: integer
        provider:
          type: string
        datasetIds:
          type: array
          items:
            type: string
        warning:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        requestId:
          $ref: "#/components/schemas/RequestId"
...
paths:
...
  /search:
    get:
      operationId: searchContent
      summary: Unified search across the MySQL search projection
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
            maximum: 50
      responses:
        "200":
          description: Search matches from public/private projections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          description: Missing or invalid API key.
  /search/vector:
    get:
      operationId: searchLiteraryVector
      summary: Vector search across the RAGFlow literary corpus projection
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
            maximum: 50
        - name: ragflowDatasetIds
          in: query
          schema:
            type: string
          description: Optional comma-separated RAGFlow dataset ids. Defaults to contentmrs-literary-corpus.
        - name: similarityThreshold
          in: query
          schema:
            type: number
            default: 0.05
        - name: vectorWeight
          in: query
          schema:
            type: number
            default: 0.35
        - name: topK
          in: query
          schema:
            type: integer
            default: 64
      responses:
        "200":
          description: Literary corpus chunks retrieved through RAGFlow vector retrieval.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "401":
          description: Missing or invalid API key.
        "502":
          description: RAGFlow retrieval failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...